General Imporvements
Added support for text being dark as well as support for buttons being dark. Unable to fix a bug that makes the button go down when there is more than one line of text, so temportary fixed it by capping text limit to one line. Changed how the navagation looks when media screen is small. Also cleaned a little bit of the code :P